ChangeFormer/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1
_openmp_mutex=4.5
blas=1.0
brotli=1.0.9
bzip2=1.0.8
ca-certificates=2021.10.8
certifi=2021.10.8
colorama=0.4.4
cudatoolkit=10.2.89
cycler=0.11.0
dbus=1.13.18
einops=0.3.2
expat=2.2.10
ffmpeg
fontconfig=2.13.1
fonttools=4.25.0
freetype=2.11.0
gettext=0.19.8.1
giflib=5.2.1
glib=2.58.3
gmp=6.2.1
gnutls=3.6.15
gst-plugins-base=1.14.5
gstreamer=1.14.5
icu=64.2
intel-openmp=2021.4.0
jpeg=9d
kiwisolver=1.3.1
lame=3.100
lcms2=2.12
libedit=3.1.20210910
libffi=3.2.1
libgcc-ng=9.3.0
libgfortran-ng=7.5.0
libgfortran4=7.5.0
libgomp=9.3.0
libiconv=1.15
libidn2=2.3.2
libpng=1.6.37
libstdcxx-ng=9.3.0
libtasn1=4.16.0
libtiff=4.2.0
libunistring=0.9.10
libuuid=2.32.1
libuv=1.40.0
libwebp=1.2.0
libwebp-base=1.2.0
libxcb=1.13
libxml2=2.9.10
lz4-c=1.9.3
matplotlib=3.4.3
matplotlib-base=3.4.3
mkl=2021.4.0
mkl-service=2.4.0
mkl_fft=1.3.1
mkl_random=1.2.2
munkres=1.1.4
ncurses=6.3
nettle=3.7.3
numpy=1.21.2
numpy-base=1.21.2
olefile=0.46
openh264=2.1.1
openssl=1.1.1l
pcre=8.45
pillow=8.4.0
pip=21.2.4
pthread-stubs=0.4
pyparsing=3.0.6
pyqt=5.9.2
python=3.8.0
python-dateutil=2.8.2
python_abi=3.8
pytorch=1.10.1
pytorch-mutex=1.0
qt=5.9.7
readline=7.0
scipy=1.7.1
setuptools=58.0.4
sip=4.19.13
six=1.16.0
sqlite=3.33.0
timm=0.4.12
tk=8.6.11
torchaudio=0.10.1
torchvision=0.11.2
tornado=6.1
tqdm=4.62.3
typing_extensions=3.10.0.2
wheel=0.37.0
xorg-libxau=1.0.9
xorg-libxdmcp=1.1.3
xz=5.2.5
zlib=1.2.11
zstd=1.4.9
